requirements.txt
# ── AI / LLM ──────────────────────────────────────────────
langgraph>=0.3.0
langchain-core>=0.3.0
langgraph-checkpoint-sqlite>=2.0.0

# ── Web Framework ───────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# ── LLM Clients ─────────────────────────────────────────────
openai>=1.60.0
httpx>=0.28.0
websockets>=13.0

# ── Browser Automation (form filling) ──────────────────────
playwright>=1.49.0
Pillow>=11.0.0

# ── Voice (TTS fallback) ─────────────────────────────────
edge-tts>=6.0

# ── Security ───────────────────────────────────────────────
cryptography>=44.0.0

# ── Environment ───────────────────────────────────────────
python-dotenv>=1.0.1
aiosqlite>=0.21.0

# ── Monitoring (optional, for production) ────────────────
prometheus-client>=0.20.0